sentinel/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel"
version = "0.1.0"
description = "File system watcher for Obsidian Vault ingestion (Bronze Tier)"
requires-python = ">=3.10"
authors = [
    {name = "Digital FTE", email = "noreply@example.com"},
]
dependencies = [
    "watchdog>=3.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
